Fix Meson introspect to not use relative paths
The previous fix correctly set the working directory for meson introspect
commands, but still passed "./meson.build" as an argument. When meson
introspect is run from the project directory (via cwd), it automatically
finds the meson.build file and doesn't need a path argument.

This was causing the --scan-dependencies call to fail or hang because
meson was looking for "./meson.build" relative to an already-set working
directory.

Removed the unnecessary "./meson.build" argument from the introspect
method since the working directory is already set to the project directory.